Move `add_metering` out of `wasmer` module
It will no longer be Wasmer specific.
