+++ linera-execution/src/wasm/mod.rs
@@ -22,6 +22,7 @@ mod wasmtime;
 
 use linera_base::data_types::Bytecode;
 use thiserror::Error;
+use wasm_instrument::{gas_metering, parity_wasm};
 #[cfg(with_wasmer)]
 use wasmer::{WasmerContractInstance, WasmerServiceInstance};
 #[cfg(with_wasmtime)]
@@ -207,6 +208,49 @@ impl UserServiceModule for WasmServiceModule {
 
         Ok(instance)
     }
+}
+
+/// Instrument the [`Bytecode`] to add fuel metering.
+pub fn add_metering(bytecode: Bytecode) -> anyhow::Result<Bytecode> {
+    struct WasmtimeRules;
+
+    impl gas_metering::Rules for WasmtimeRules {
+        /// Calculates the fuel cost of a WebAssembly [`Operator`].
+        ///
+        /// The rules try to follow the hardcoded [rules in the Wasmtime runtime
+        /// engine](https://docs.rs/wasmtime/5.0.0/wasmtime/struct.Store.html#method.add_fuel).
+        fn instruction_cost(
+            &self,
+            instruction: &parity_wasm::elements::Instruction,
+        ) -> Option<u32> {
+            use parity_wasm::elements::Instruction::*;
+
+            Some(match instruction {
+                Nop | Drop | Block(_) | Loop(_) | Unreachable | Else | End => 0,
+                _ => 1,
+            })
+        }
+
+        fn memory_grow_cost(&self) -> gas_metering::MemoryGrowCost {
+            gas_metering::MemoryGrowCost::Free
+        }
+
+        fn call_per_local_cost(&self) -> u32 {
+            0
+        }
+    }
+
+    let instrumented_module = gas_metering::inject(
+        parity_wasm::deserialize_buffer(&bytecode.bytes)?,
+        gas_metering::host_function::Injector::new(
+            "linera:app/contract-runtime-api",
+            "consume-fuel",
+        ),
+        &WasmtimeRules,
+    )
+    .map_err(|_| anyhow::anyhow!("failed to instrument module"))?;
+
+    Ok(Bytecode::new(instrumented_module.into_bytes()?))
 }
 
 #[cfg(web)]
